:truck: Move list store service
