+++ src/app/list-add/list-add.component.ts
@@ -4,7 +4,7 @@ import { MatSnackBar } from '@angular/material';
 
 import { GoogleAnalyticsService } from '../google-analytics.service';
 import { ListService } from '../list.service';
-import { ListStoreService } from '../list-store.service';
+import { ListStoreService } from '../services/store/list-store.service';
 
 import { List } from '../models/list';
