+++ src/app/services/store/list-store.service.spec.ts
