+++ src/app/sidebar/sidebar.component.ts
@@ -7,7 +7,7 @@ import { List } from '../models/list';
 
 import { AuthService } from '../auth.service';
 import { ListService } from '../list.service';
-import { ListStoreService } from '../list-store.service';
+import { ListStoreService } from '../services/store/list-store.service';
 import { SidebarService } from '../sidebar.service';
 
 @Component({
